refactor(web): extract WhatsApp onboarding into web_routers/whatsapp_onboarding (web_server.py god-file slice R4-C3)
